Adicionado um menu para selecionar 2 funcionalidade do programa principal
